+++ system/src/models/items/AncestrySD.mjs
@@ -11,6 +11,7 @@ export default class AncestrySD extends BaseItemSD {
 			talentChoiceCount: new fields.NumberField({ integer: true, initial: 1, min: 0}),
 			nameTable: new fields.DocumentUUIDField(),
 			nameTable2: new fields.DocumentUUIDField(),
+			trinketTable: new fields.DocumentUUIDField(),
 			randomWeight: new fields.NumberField({ integer: true, initial: 1, min: 1}),
 		};